requirements.txt
Django==1.7.1
Fabric==1.10.0
django-bower==5.0.1
django-filter==0.8
django-filters==0.1.0
django-password-reset==0.7
djangorestframework==2.4.4
drf-nested-routers==0.9.0
ecdsa==0.11
fabric-virtualenv==0.2.1
paramiko==1.15.1
pycrypto==2.6.1
six==1.8.0
wsgiref==0.1.2
